Unify action prediction with decision making under the umbrella of strategy.
A strategy is a function that accepts a state and returns a probability distribution of actions. When used for prediction, it's the probability that the foe uses each action assuming they are following that strategy. When used for decision making, the probabilities represent the distribution that TM should use to select from those moves.

`score_selections` is now the `expectimax` strategy.